requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
channels==4.0.0
channels-redis==4.1.0
Pillow==10.1.0
requests==2.31.0
# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
gunicorn==21.2.0
daphne==4.0.0
whitenoise==6.6.0
django-health-check==3.17.0
sentry-sdk==1.38.0